DataScience/Jupyter Notebook/FireDetect/3DFire/fire-loc3d-cgt/requirements.txt
ultralytics>=8.2.0
torch>=2.1.0
torchvision>=0.16.0
opencv-contrib-python>=4.8.1
numpy>=1.24
scipy>=1.11
tqdm>=4.66
pyyaml>=6.0
filterpy>=1.4.5
matplotlib>=3.8
